more generic node types
