+++ crates/ir/src/graph/node.rs
@@ -1,9 +1,8 @@
-pub mod assign;
-pub mod branch;
-pub mod call;
-pub mod func;
-pub mod term;
-use std::any::Any;
+mod assign;
+mod branch;
+mod call;
+mod func;
+mod term;
 
 pub use assign::*;
 pub use branch::*;
@@ -12,6 +11,7 @@ pub use func::*;
 pub use term::*;
 
 use crate::expr::{Expr, VarExpr};
+use std::any::Any;
 
 pub trait DataFlow: dyn_clone::DynClone {
     fn referenced_vars(&self) -> Vec<&VarExpr> {
@@ -38,35 +38,6 @@ pub trait DataFlow: dyn_clone::DynClone {
     /// Return true if successful, false otherwise
     fn unreference_var(&mut self, _var: &VarExpr) -> bool {
         false
-    }
-}
-
-#[derive(Clone, PartialEq)]
-pub enum Node {
-    Assign(AssignNode),
-    Branch(BranchNode),
-    Yield(TermNode),
-    Return(TermNode),
-    Func(FuncNode),
-    Call(CallNode),
-}
-
-impl std::fmt::Display for Node {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        match self {
-            Node::Assign(n) => write!(f, "{}", n),
-            Node::Branch(n) => write!(f, "{}", n),
-            Node::Yield(n) => write!(f, "yield{}", n),
-            Node::Return(n) => write!(f, "return{}", n),
-            Node::Func(n) => write!(f, "{}", n),
-            Node::Call(n) => write!(f, "{}", n),
-        }
-    }
-}
-
-impl std::fmt::Debug for Node {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        write!(f, "{}", self)
     }
 }
